Update Theatre Square.cpp

+++ Theatre Square.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main(){
 	
-	long m,n ,a; scanf("%d %d %d",&m, &n, &a); 
+	long long m,n ,a; scanf("%d %d %d",&m, &n, &a); 
 	printf("%d\n", ((m+a-1)/a)*((n+a-1)/a)) ;
 
 	return 0;
